// src/app/pages/outgoingShippingModules/shipping-guides/interfaces/guides-ent-pop.model.ts
import { ShippingEnt } from "../../shipping-registry/interfaces/shipping.model";
import { Customer } from "../../../warehousingModules/customers-registry/interfaces/customer.model";
import { PackageType } from "../../../adminModules/package-type-registry/interfaces/package-type.model";
import { WarehouseItemFull } from "../../../warehousingModules/warehouse-inventory/interfaces/warehouse-item-full.model";
import { User } from "../../../adminModules/users-registry/interfaces/users.model";
import { Address } from "../../../warehousingModules/customers-registry/interfaces/address.model";
import { Rate } from "../../../adminModules/rate-registry/interfaces/rate.model";
import { PaymentBitacoraInterface } from "../../../../interfaces/payment-bitacora-data-table.interface";

export class GuidesEntPop {
  _id: number;
  tlCargoId: string;
  name: string;
  type: string;
  status: string;
  finalVolume: number;
  finalVlb: number;
  finalWeight: number;
  cost: number;
  creationDate: Date;
  deliveryDate: Date;
  packageList: WarehouseItemFull[];
  shipping: ShippingEnt;
  customer: Customer;
  packageType: PackageType;
  notes: string;
  imageUrl: string;
  deliveryAddress: Address;
  rateAmount: number;
  rate: Rate;
  packageTypeSelected: boolean;
  notifiedTimes: number;
  user: User;
  paymentStatus: string;
  paidAmount: number;
  paymentBitacora: PaymentBitacoraInterface[];
  notification: boolean;
  partialPaid: boolean;
  quote: boolean;

  constructor(guidesModel) {
    this._id = guidesModel._id;
    this.tlCargoId = guidesModel.tlCargoId;
    this.name = guidesModel.name;
    this.type = guidesModel.type;
    this.status = guidesModel.status;
    this.finalVolume = guidesModel.finalVolume;
    this.finalVlb = guidesModel.finalVlb;
    this.finalWeight = guidesModel.finalWeight;
    this.cost = guidesModel.cost;
    this.creationDate = guidesModel.creationDate;
    this.deliveryDate = guidesModel.deliveryDate;
    this.packageList = guidesModel.packageList;
    this.shipping = guidesModel.shipping;
    this.customer = guidesModel.customer;
    this.packageType = guidesModel.packageType;
    this.notes = guidesModel.notes;
    this.imageUrl = guidesModel.imageUrl;
    this.deliveryAddress = guidesModel.deliveryAddress;
    this.rateAmount = guidesModel.rateAmount;
    this.rate = guidesModel.rate;
    this.packageTypeSelected = guidesModel.packageTypeSelected;
    this.notifiedTimes = guidesModel.notifiedTimes;
    this.user = guidesModel.user;
    this.paymentStatus = guidesModel.paymentStatus;
    this.paidAmount = guidesModel.paidAmount;
    this.paymentBitacora = guidesModel.paymentBitacora;
    this.notification = guidesModel.notification;
    this.partialPaid = guidesModel.partialPaid;
    this.quote = guidesModel.quote;
  }
}